// src/app/shared/util-model/enumeration/element-action.enum.ts
export enum ElementActionEnum {
    USER_UPDATE = 'USER_UPDATE',
    USER_BLOCK = 'USER_BLOCK',
    USER_UNBLOCK = 'USER_UNBLOCK',
    USER_IMPERSONATE = 'USER_IMPERSONATE',
    USER_DELETE = 'USER_DELETE',

    PROJECT_SELECT_PROFILE = 'PROJECT_SELECT_PROFILE',
    PROJECT_CREATE_SUPPORT_PROFILE = 'PROJECT_CREATE_SUPPORT_PROFILE',
    PROJECT_UPDATE = 'PROJECT_UPDATE',
    PROJECT_DISABLE = 'PROJECT_DISABLE',
    PROJECT_ENABLE = 'PROJECT_ENABLE',
    PROJECT_DELETE = 'PROJECT_DELETE',

    PROJECT_PROFILE_SELECT = 'PROJECT_PROFILE_SELECT',
    PROJECT_PROFILE_UPDATE = 'PROJECT_PROFILE_UPDATE',
    PROJECT_PROFILE_BLOCK = 'PROJECT_PROFILE_BLOCK',
    PROJECT_PROFILE_UNBLOCK = 'PROJECT_PROFILE_UNBLOCK',
    PROJECT_PROFILE_DELETE = 'PROJECT_PROFILE_DELETE',

    PARTICIPANT_CONSULT_MOVEMENTS = 'PARTICIPANT_CONSULT_MOVEMENTS',
    PARTICIPANT_UPDATE = 'PARTICIPANT_UPDATE',
    PARTICIPANT_DISABLE = 'PARTICIPANT_DISABLE',
    PARTICIPANT_ENABLE = 'PARTICIPANT_ENABLE',
    PARTICIPANT_REMOVE_FROM_GROUP = 'PARTICIPANT_REMOVE_FROM_GROUP',
    PARTICIPANT_DELETE = 'PARTICIPANT_DELETE',

    MOVEMENT_CONSULT_COMMUNICATIONS = 'MOVEMENT_CONSULT_COMMUNICATIONS',
    MOVEMENT_UPDATE = 'MOVEMENT_UPDATE',
    MOVEMENT_DISABLE = 'MOVEMENT_DISABLE',
    MOVEMENT_ENABLE = 'MOVEMENT_ENABLE',
    MOVEMENT_DELETE = 'MOVEMENT_DELETE',

    GROUP_CONSULT_MEMBERS = 'GROUP_CONSULT_MEMBERS',
    GROUP_UPDATE = 'GROUP_UPDATE',
    GROUP_DISABLE = 'GROUP_DISABLE',
    GROUP_ENABLE = 'GROUP_ENABLE',
    GROUP_DELETE = 'GROUP_DELETE',

    VEHICLE_CONSULT_MOVEMENTS = 'VEHICLE_CONSULT_MOVEMENTS',
    VEHICLE_UPDATE = 'VEHICLE_UPDATE',
    VEHICLE_DISABLE = 'VEHICLE_DISABLE',
    VEHICLE_ENABLE = 'VEHICLE_ENABLE',
    VEHICLE_DELETE = 'VEHICLE_DELETE',

    ACTIVITY_CONSULT_MOVEMENTS = 'ACTIVITY_CONSULT_MOVEMENTS',
    ACTIVITY_UPDATE = 'ACTIVITY_UPDATE',
    ACTIVITY_DISABLE = 'ACTIVITY_DISABLE',
    ACTIVITY_ENABLE = 'ACTIVITY_ENABLE',
    ACTIVITY_DELETE = 'ACTIVITY_DELETE',

    COMMUNICATION_UPDATE = 'COMMUNICATION_UPDATE',
    COMMUNICATION_DISABLE = 'COMMUNICATION_DISABLE',
    COMMUNICATION_ENABLE = 'COMMUNICATION_ENABLE',
    COMMUNICATION_DELETE = 'COMMUNICATION_DELETE',
}